Advanced caching for resolve and entity collection

+++ apimodel/collection.go
@@ -6,7 +6,7 @@ package apimodel
 // language tags according to BCP47 (RFC5646).
 type EntityCollectionRequest struct {
 	FromEntityID   string   `json:"from_entity_id" form:"from_entity_id" query:"from_entity_id" url:"from_entity_id"`
-	Limit          uint64   `json:"limit" form:"limit" query:"limit" url:"limit"`
+	Limit          int      `json:"limit" form:"limit" query:"limit" url:"limit"`
 	EntityTypes    []string `json:"entity_type" form:"entity_type" query:"entity_type" url:"entity_type"`
 	TrustMarkTypes []string `json:"trust_mark_type" form:"trust_mark_type" query:"trust_mark_type" url:"trust_mark_type"`
 	TrustAnchor    string   `json:"trust_anchor" form:"trust_anchor" query:"trust_anchor" url:"trust_anchor"`
